requirements.txt
fastapi
fastapi-cli
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pytest
numpy
pandas
requests
langchain
langchain-openai
langchain-community
python-dotenv
asyncpg
